+++ examples/mocha-multi-library/lib/Wall.js
@@ -1,0 +1,11 @@
+var Wall = function(items) {
+    this.items = items;
+    this.fall = function(n) {
+        this.items -= n;
+    }
+    this.returned = function() {
+        this.items++;
+    }
+}
+
+module.exports = Wall;