+++ examples/mocha-multi-library/lib/dictionary.js
@@ -1,0 +1,7 @@
+var Yadda = require('yadda');
+var Dictionary = Yadda.Dictionary;
+
+module.exports = (function() {
+    return new Dictionary()
+        .define('NUM', /(\d+)/);
+})();